requirements.txt
locust
pandas
streamlit
python-dotenv